divide off tree commands from element commands
